DR_FAN/requirements.txt
torch
matplotlib
tqdm
shapely
numpy
pillow
functools
psutil
pyparsing
